requirements.txt
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
optuna>=2.10.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0
PyYAML>=5.4.1
imbalanced-learn>=0.8.0
